Resolving issues

+++ src/kaur/siimar/client/FileDetails.java
@@ -5,16 +5,16 @@ public class FileDetails {
     //once we have input, we save all these into FileDetails class.
 
     private String path;
-    private int size;
+    private long size;
     private String content;
 
-    public FileDetails(String path, int size, String content) {
+    public FileDetails(String path, long size, String content) {
         this.path = path;
         this.size = size;
         this.content = content;
     }
 
-    public int getSize() {
+    public long getSize() {
         return this.size; // cannot access it beyond this class because private attribute
     }
